requirements.txt
uvicorn==0.27.1
fastapi==0.110.0
pydantic==2.6.2
cohere==4.51
faiss-cpu==1.7.4
scikit-learn==1.4.1.post1
openai==1.12.0
python-dotenv==1.0.1
redis==5.0.1